Updates PyPi descriptions
Fixes Update PyPi description and README #3

Ideally we want this to be the README.md of the `findent` project.
However, for the time being it is slightly outdated and format-wise it
does not adhere to Markdown so we make our own copy.

TODO: If/when findent updates their README.md discart this file