Added automatic versioning based on tags
The versioning does not work on powershell so we have to use bash.
Moreover, we are defining the env: variables at the global scope
by writting in GITHUB_ENV i.e. to disk. These is suboptimal and cryptic
IMO but there is no other way to set global variables in Actions.