build: switching to scikit-build and cmake
Uses the Unix Makefiles as a build tool since for Windows
the MSVC generator is overrides the compiler to MSVC instead of g++.
Similarly, Ninja was causing some issues again defaulting to MSVC
compilers.

For Linux and MacOS, we use flex and bison in the cibuildwheel setup
but for Windows we cannot rely on that so we use the default settings.

The `assets` folder was moved into `docs` because it was interfering
with the flat layout of the build directory and scikit-build.

The version for both cmake and setuptools (scikit-build) is read
from a text file. Now only this file needs to be edited to update
the version of findent being build and packaged.